# openapi.yaml
openapi: 3.0.3
info:
  title: ERP/POS API
  description: |
    Sistema ERP de punto de venta y gestión de almacén.
    
    ## Características principales
    - Autenticación JWT con roles
    - Gestión de productos con variantes
    - Control de inventario en tiempo real
    - Punto de venta con precios automáticos (mayoreo/menudeo)
    - Facturación SAT preparada
    - Multi-caja y multi-usuario
    
    ## Arquitectura
    - Backend: FastAPI + PostgreSQL
    - Frontend: React
    - Base de datos con lógica de negocio centralizada
    
  version: 1.0.0
  contact:
    name: Soporte Técnico
    email: soporte@erp.com
servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
  - url: https://api.erp.com/api/v1
    description: Producción
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido del endpoint /auth/login.
        Incluye en el header: `Authorization: Bearer {token}`

  # ============================================
  # ESQUEMAS
  # ============================================
  schemas:
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "STOCK_INSUFFICIENT"
        message:
          type: string
          example: "Stock insuficiente para completar la operación"
        details:
          type: object
          additionalProperties: true

    # --- Auth ---
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: "admin"
        password:
          type: string
          format: password
          example: "admin123"
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          example: 3600
        user:
          $ref: '#/components/schemas/UserInfo'
    UserInfo:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "admin"
        email:
          type: string
          example: "admin@erp.com"
        rol:
          type: string
          enum: [ADMIN, CAJERO]
          example: "ADMIN"
        punto_venta_id:
          type: integer
          nullable: true
          example: 1

    # --- Productos y Variantes ---
...
    FacturaCreateRequest:
      type: object
      required:
        - ventas_ids
        - rfc_receptor
        - uso_cfdi
      properties:
        ventas_ids:
          type: array
          items:
            type: integer
          example: [101, 102]
        rfc_receptor:
          type: string
          example: "XAXX010101000"
        uso_cfdi:
          type: string
          example: "G03"
        nombre_receptor:
          type: string
          example: "PUBLICO EN GENERAL"
        tipo_factura:
          type: string
          enum: [INDIVIDUAL, GLOBAL]
          example: "INDIVIDUAL"
    FacturaResponse:
      type: object
      properties:
        factura_id:
          type: integer
        uuid:
          type: string
          nullable: true
        folio:
          type: string
        serie:
          type: string
        estado:
          type: string
          enum: [BORRADOR, TIMBRADA, CANCELADA]
        total:
          type: number
        xml_url:
          type: string
          nullable: true
        pdf_url:
          type: string
          nullable: true

# ============================================
# ENDPOINTS
# ============================================
paths:
...
  /auth/login:
    post:
      tags:
        - Autenticación
      summary: Iniciar sesión
      description: |
        Autentica un usuario y devuelve un token JWT.
        El token debe incluirse en todas las peticiones protegidas.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      tags:
        - Autenticación
      summary: Obtener información del usuario actual
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Información del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: No autenticado

  # ============================================
  # POS - FLUJO CRÍTICO
  # ============================================
...
  /facturas:
    post:
      tags:
        - Facturación
      summary: Crear factura (borrador)
      description: |
        Crea una factura a partir de una o varias ventas.
        Estado inicial: BORRADOR.
        Para timbrar, usar POST /facturas/{id}/timbrar
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FacturaCreateRequest'
      responses:
        '201':
          description: Factura creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacturaResponse'
  /facturas/{factura_id}:
    get:
      tags:
        - Facturación
      summary: Obtener detalle de factura
      security:
        - BearerAuth: []
      parameters:
        - name: factura_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalle de factura
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacturaResponse'
  /facturas/{factura_id}/timbrar:
    post:
      tags:
        - Facturación
      summary: Timbrar factura con PAC
      description: |
        Envía la factura al PAC para obtener el UUID y el XML/PDF timbrado.
        Solo aplica a facturas en estado BORRADOR.
      security:
        - BearerAuth: []
      parameters:
        - name: factura_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Factura timbrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacturaResponse'
        '409':
          description: Factura ya timbrada o error del PAC
  /facturas/global/tarjetas:
    post:
      tags:
        - Facturación
      summary: Generar factura global de ventas con tarjeta
      description: |
        Genera una factura global agrupando todas las ventas con método de pago TARJETA
        en un rango de fechas (típicamente día completo).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - fecha_desde
                - fecha_hasta
              properties:
                fecha_desde:
                  type: string
                  format: date
                  example: "2026-01-14"
                fecha_hasta:
                  type: string
                  format: date
                  example: "2026-01-14"
                punto_venta_id:
                  type: integer
                  nullable: true
      responses:
        '201':
          description: Factura global creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacturaResponse'

  # ============================================
  # USUARIOS (ADMIN)
  # ============================================
...